+++ src/index.html
@@ -129,7 +129,9 @@
 
 <script src="js/form.js"></script>
 <script src="js/game.js"></script>
-<script src="reviews.js"></script>
+<script src="js/load.js"></script>
+<script src="js/review.js"></script>
+<script src="js/reviews.js"></script>
 <script src="js/main.js"></script>
 </body>
 </html>
